Style: streamline interpolation expressions
Only use "interpolate" expressions instead of "stops" or "type":
"interval" functions. Remove useless interpolation with "zoom" 0, which
was generated by Maputnik. If there is only one filter expression remove
the "all" brace.

The goal is to use less features of the style spec so the gap what would
be needed to supported in JOSM gets smaller.