Use filters for colors instead of separate layers
JOSM doesn't support accessing different layers in the MVT.

But it turns out that it supports simple filter expressions. There can
be only one filter and it has to be of the form

    "filter": ["==", "category", 0]

and not what Maputnik generates:

    "filter": ["all", ["==", "category", 0]]

So use filter instead to assign the colors.

There were some problems with polygons not being displayed in JOSM. I
suspect a polygon winding order problem, but couldn't nail it down, yet.
So throwing some ST_ForcePolygonCW into the mix to hopefully make sure
that the winding order is consistent.