Move detailed view one zoom level down
Move hexagons zoom 0--12 --> 0--13
Points 13 --> 14
Details 14 --> 15

Increase the maxzoom of the style to 15.

Change the mapping from tile zoom to H3 resolution to be zoom-1.

JOSM's style conversion can't handle multiple curly brace placeholders
in text labels. So concatenate the label already in the MVTs, and in the
style only use one field.

When generating the classes for the colors, there was a logic error when
using evenly spaced start classes for k-means. Fix it by always
generating 8 start classes.