Add a Mapbox GL Style compliant style
Mapbox GL Styles use continuous zoom intervals where the minzoom is
inclusive and the maxzoom is exclusive. When one layer is defined with
minzoom 12 to maxzoom 14 and another one with minzoom 14 to maxzoom 16,
they a valid for 12.0 to 13.99 and from 14.0 to 15.99, respectively.

In JOSM on the other hand, the maxzoom is taken +1. I.e. minzoom 12 to
maxzoom 14 would be valid from 12.0 to 14.99.

To keep the main style.json JOSM-compatible, it uses exclusive
intervals: 0-13, 14-14, 15-15, etc. Add a variant mapbox-gl-style.json
which uses exclusive intervals 0-14, 14-15, 15-16, etc. This way the
style can be used with Mapbox GL, or can be loaded in Maputnik.