Zoom bounds, URL hash for map view
Correctly set the map bounding box. This helps so viewers can't get lost
in other parts of the world where they can't find back to the area with
data.

Sync the map position with a URL fragment with the zoom and center
coordinate. This is nice because it allows you to link to a certain map
view.