Use continuous zoom intervals
A bug in JOSM was fixed in https://josm.openstreetmap.de/ticket/22404 to
correctly render polygons in MVTs. Along the way an issue with the zoom
levels was fixed: Previously, the zoom levels had to be exclusive, i.e.
0-13, 14-14, 15-15 etc.

With the fix in place, the zoom levels are now continuous -- the same as
in Mapbox GL styles. We can now use 0-14, 14-15, 15-16, etc.

This way style.json becomes the same as mapbox-gl-style.json and makes
it redundant. But as the file is linked in the discussion of above bug
report, just keep it around.

See also 3361f5e86b984b7deac3a03850a4fbbc7f316920