Ignore allotments
Buildings in only a few of all the allotments are listed with "official"
house numbers. As the name of the allotment is used as street name, and
it's unclear whether these house numbers are really used on the ground,
they should not be added to OpenStreetMap like this.

To avoid the wrong impression that these addresses are somehow
"missing", don't draw them as pink in the style but as grey instead.