Allow NULL values in address parts
LGB has changed the format of the address part columns which can now
contain NULL. This resulted in the naively concatenated address string
to be NULL as well, and thus addresses missing from MVT tiles where e.g.
the suburb is NULL.

The address string generation was made more robust. Also omit the suburb
if it's equal to the municipality name.