Separate numerical house number extension with /
House numbers can have extensions, usually a, b, c. But there are some
house numbers with numerical extensions 1, 2, 3.

In the LGB data, 'hnr' contains the house number, 'adz' the optional
extension (or an empty string) and 'hnradz' is supposed to contain the
already merged housenumber with extension.

'hnradz' contains the simple string concatenation of 'hnr' and 'adz',
which is wrong if 'adz' starts with a digit 0-9.

If 'adz' starts with a digit, separate it with a slash /. This results
in house numbers 1/1, 1/2, ... which seems to be the convention in OSM.

Rename the new column to 'house_number' to avoid confusion with 'hnradz'
from the raw data.