Split plot numbers with '/'
There are "combined house numbers" with plot numbers which contain a
house number for the larger property and then an additional plot number
for each smaller plot inside the property. The numbers can look like
5/p1, 5/p2, ... or even 5a/p1, 5a/p2, ...

Note that there are also house numbers of the form number-letter-number
such as "5a1" where no extra slash should be inserted.