Breakpoints: Use percentiles for all resolutions
kmeans is used to calculate breakpoints for coloring the results. As
input kmeans expects an array of start values. There was a problem in
the previous implementation where generate_series() was used, but could
generate the wrong number of values. In this case, kmeans would abort
with an exception, which aborted the whole process.

Instead, always use percentiles for all resolutions. For high
resolutions, so far there was the problem that "1" appeared so often it
would appear in several percentiles. kmeans should not be given several
identical starting values, because then it returns too few classes. To
solve this, make the list of start values distinct and strictly
increasing.